### 2023-06-20 #5574 @pswpswpsw
Hi Daniel,

Sorry for the trouble. I generate PDF from my end with the following
command.

 pandoc -s --bibliography paper.bib --citeproc paper.md --csl default.csl
-o paper.pdf

I will go back to cut the reference bib.

Best,

==========
Shaowu Pan
Assistant Professor of Aerospace Engineering
Rensselaer Polytechnic Institute

On Tue, Jun 20, 2023 at 2:17 PM Daniel S. Katz ***@***.***>
wrote:

> I'm also not sure why your paper is not generating, so I'll try the
> command here.
>
