+++ docs/JOSS/paper.bib
@@ -1,690 +1,8 @@
 @IEEEtranBSTCTL{IEEEexample:BSTcontrol,
   CTLdash_repeated_names = "no"
 }
-@article{brunton2016pnas,
-  title = {Discovering governing equations from data by sparse identification of nonlinear dynamical systems},
-  author = {Brunton, Steven L. and Proctor, Joshua L. and Kutz, J. Nathan},
-  journal = {Proceedings of the National Academy of Sciences},
-  volume = {113},
-  number = {15},
-  pages = {3932--3937},
-  year = {2016},
-  publisher = {National Academy of Sciences},
-  doi = {10.1073/pnas.1517384113}
-}
-
-@article{lange2021fourier,
-  title={From {F}ourier to {K}oopman: Spectral methods for long-term time series prediction},
-  author={Lange, Henning and Brunton, Steven L and Kutz, J Nathan},
-  journal={The Journal of Machine Learning Research},
-  volume={22},
-  number={1},
-  pages={1881--1918},
-  year={2021},
-  publisher={JMLRORG}
-}
-@article{xiong2023koopmanlab,
-  title={Koopman{L}ab: A PyTorch module of {K}oopman neural operator family for solving partial differential equations},
-  author={Xiong, Wei and Ma, Muyuan and Sun, Pei and Tian, Yang},
-  journal={arXiv preprint arXiv:2301.01104},
-  year={2023}
-}
-@article{abraham2019active,
-  title={Active learning of dynamics for data-driven control using {K}oopman operators},
-  author={Abraham, Ian and Murphey, Todd D},
-  journal={IEEE Transactions on Robotics},
-  volume={35},
-  number={5},
-  pages={1071--1083},
-  year={2019},
-  publisher={IEEE}
-}
-@article{mamakoukas2021derivative,
-  title={Derivative-based {K}oopman operators for real-time control of robotic systems},
-  author={Mamakoukas, Giorgos and Castano, Maria L and Tan, Xiaobo and Murphey, Todd D},
-  journal={IEEE Transactions on Robotics},
-  volume={37},
-  number={6},
-  pages={2173--2192},
-  year={2021},
-  publisher={IEEE}
-}
-@article{pan2020structure,
-  title={On the structure of time-delay embedding in linear models of non-linear dynamical systems},
-  author={Pan, Shaowu and Duraisamy, Karthik},
-  journal={Chaos: An Interdisciplinary Journal of Nonlinear Science},
-  volume={30},
-  number={7},
-  pages={073135},
-  year={2020},
-  publisher={AIP Publishing LLC}
-}
-@article{kingma2014adam,
-  title={Adam: A method for stochastic optimization},
-  author={Kingma, Diederik P and Ba, Jimmy},
-  journal={arXiv preprint arXiv:1412.6980},
-  year={2014}
-}
-@book{nocedal1999numerical,
-  title={Numerical optimization},
-  author={Nocedal, Jorge and Wright, Stephen J},
-  year={1999},
-  publisher={Springer}
-}
-@article{wright1999numerical,
-  title={Numerical optimization},
-  author={Wright, Stephen and Nocedal, Jorge and others},
-  journal={Springer Science},
-  volume={35},
-  number={67-68},
-  pages={7},
-  year={1999}
-}
-@article{ljung2010arc,
-  title={Perspectives on system identification},
-  author={Ljung, Lennart},
-  journal={Annual Reviews in Control},
-  volume={34},
-  number={1},
-  pages={1--12},
-  year={2010},
-  publisher={Elsevier}
-}
-@incollection{Nelles2013book,
-  title={Nonlinear Dynamic System Identification},
-  author={Nelles, Oliver},
-  booktitle={Nonlinear System Identification},
-  pages={831--891},
[…]
