+++ docs/JOSS/paper.bib
@@ -10,12 +10,14 @@
   volume={656},
   pages={5--28},
   year={2010},
+  doi={10.1017/S0022112010001217},
   publisher={Cambridge University Press}
 }
 @book{Kutz2016book,
   title={Dynamic mode decomposition: data-driven modeling of complex systems},
   author={Kutz, J Nathan and Brunton, Steven L and Brunton, Bingni W and Proctor, Joshua L},
   year={2016},
+  doi={10.1137/1.9781611974508},
   publisher={SIAM}
 }
 
@@ -26,9 +28,9 @@
 	pages = {225--254},
 	title = {Dynamic mode decomposition and its variants},
 	volume = {54},
+  doi={10.1146/annurev-fluid-030121-015835},
 	year = {2022}
 }
-
 
 
 @article{ljung2010arc,
@@ -39,6 +41,7 @@
   number={1},
   pages={1--12},
   year={2010},
+  doi={10.1016/j.arcontrol.2009.12.001},
   publisher={Elsevier}
 }
 @article{wright1999numerical,
@@ -48,12 +51,14 @@
   volume={35},
   number={67-68},
   pages={7},
+  doi={10.1007/978-0-387-40065-5},
   year={1999}
 }
 @book{Brunton2019book,
   title={Data-driven science and engineering: Machine learning, dynamical systems, and control},
   author={{Brunton}, Steven L. and Kutz, J Nathan},
   year={2022},
+  doi={10.1017/9781108380690},
   publisher={Cambridge University Press}
 }
 
@@ -76,6 +81,7 @@ doi = {10.1137/21M1401243}
   number={4},
   pages={047510},
   year={2012},
+  doi={10.1063/1.4772195},
   publisher={American Institute of Physics}
 }
 @article{klus2017data,
@@ -86,6 +92,7 @@ doi = {10.1137/21M1401243}
   number={3},
   pages={985--1010},
   year={2018},
+  doi={10.1007/s00332-017-9437-7},
   publisher={Springer}
 }
 
@@ -96,6 +103,7 @@ doi = {10.1137/21M1401243}
   volume={45},
   pages={357--378},
   year={2013},
+  doi={10.1146/annurev-fluid-011212-140652},
   publisher={Annual Reviews}
 }
 @article{Li2017chaos,
@@ -104,7 +112,8 @@ doi = {10.1137/21M1401243}
   journal={Chaos: An Interdisciplinary Journal of Nonlinear Science},
   volume={27},
   number={10},
-  pages={103111},
+  pages={103--111},
+  doi={10.1063/1.4993854},
   year={2017},
   publisher={AIP Publishing LLC}
 }
@@ -116,6 +125,7 @@ doi = {10.1137/21M1401243}
   volume={21},
   number={1},
   pages={243--247},
+  doi={10.1007/BF02532251},
   year={1969},
   publisher={Springer}
 }
@@ -126,6 +136,7 @@ doi = {10.1137/21M1401243}
   volume={8},
   number={1},
   pages={1--9},
+  doi={10.1038/s41467-017-00030-8},
   year={2017},
   publisher={Nature Publishing Group}
 }
@@ -135,6 +146,7 @@ doi = {10.1137/21M1401243}
   booktitle={Nonlinear System Identification},
   pages={831--891},
   year={2020},
+  doi={10.1007/978-3-662-04323-3_15},
   publisher={Springer}
 }
 @article{rowley2009spectral,
@@ -144,12 +156,14 @@ doi = {10.1137/21M1401243}
   volume={641},
   pages={115--127},
   year={2009},
+  doi={10.1017/S0022112009992059},
   publisher={Cambridge University Press}
 }
 @book{Billings2013book,
[…]
