+++ README.rst
@@ -53,7 +53,7 @@ Features implemented
 
   - Identity (for DMD/DMDc or in case users want to compute observables themselves):
     ``Identity``
-  - Multivariate polynomials: ```Polynomial``
+  - Multivariate polynomials: ``Polynomial``
   - Time delay coordinates: ``TimeDelay```
   - Radial basis functions: ``RadialBasisFunctions``
   - Random Fourier features: ``RandomFourierFeatures``
