### 2023-06-24 1ef9326
change default behavior of dmdc for svd_output_rank
