### 2023-06-25 f8a9dfb
bug in readme.rst
