### 2023-09-21 #5574 @olexandr-konovalov
@fandreuz @ndem0 @sichinaga @mtezzele I am contacting you as main developers of PyDMD - would one of you be willing and able to review this submission for JOSS, please? If you haven't reviewed for JOSS before, here we have the [Reviewer Guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
