### 2023-09-22 c77b273
fixed some building issues
