### 2023-09-26 #5881 @editorialbot
[REVIEW]: PyKoopman: A Python Package for Data-Driven Approximation of the Koopman Operator
**Submitting author:** @pswpswpsw (<a href="http://orcid.org/0000-0002-2462-362X">Shaowu Pan</a>)
**Repository:** https://github.com/dynamicslab/pykoopman
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
