### 2023-09-30 #5881 @fandreuz
Hi @pswpswpsw, a few comments about the paper:
- Figures are not really integrated in the body
   + `Figure 3` should be shrinked and it would be preferrable to move it closer to where it is referenced, I've found it unpleasant to move up and down to understand what you were talking about.
   + I feel `Figure 1` is quite detached from the context of the paper. It seems to me that it's neither referenced nor explained in the body, and it is too dense of information.
   + I had similar feelings for `Figure 2` as for `Figure 3`, I would shrink it and make it better integrated with the body.
- Please spend some more words on your dependencies, since you are pulling in some very dense ones which are not just "infrastructure". As a user I would like to know how and where they are used. For instance, I see that you have your own implementation of all DMD variants: I think this is worth noting in the paper, since you depend on PyDMD, and it's also interesting to explain why you do so instead of relying on the PyDMD implementation.
- I see you have many submodules which provide various functionalities, but I would like to see at least one code sample for one meaningful usecase, as stated in the requirements for the paper.
   + It would also be interesting if you could spend some words on what is the typical pipeline for a usage of your package. In the paper you explained the role of each submodule, but it is difficult to grasp the big picture.
- As part of the statement of need, I'd like to see a comparison with current open source tools like PyDMD (regression), SciPy (differentation). It can be one sentence, what you feel was missing and what you improved as part of your work. This is part of paper requirements (see **State of the field**).
