+++ docs/JOSS/paper.bib
@@ -54,6 +54,34 @@
   doi={10.1007/978-0-387-40065-5},
   year={1999}
 }
+@article{paszke2019pytorch,
+  title={Pytorch: An imperative style, high-performance deep learning library},
+  author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
+  journal={Advances in neural information processing systems},
+  volume={32},
+  year={2019}
+}
+@article{demo2018pydmd,
+  title={PyDMD: Python dynamic mode decomposition},
+  author={Demo, Nicola and Tezzele, Marco and Rozza, Gianluigi},
+  journal={Journal of Open Source Software},
+  volume={3},
+  number={22},
+  pages={530},
+  year={2018}
+}
+@article{kaptanoglu2022pysindy,
+  doi = {10.21105/joss.03994},
+  url = {https://doi.org/10.21105/joss.03994},
+  year = {2022},
+  publisher = {The Open Journal},
+  volume = {7},
+  number = {69},
+  pages = {3994},
+  author = {Alan A. Kaptanoglu and Brian M. de Silva and Urban Fasel and Kadierdan Kaheman and Andy J. Goldschmidt and Jared Callaham and Charles B. Delahunt and Zachary G. Nicolaou and Kathleen Champion and Jean-Christophe Loiseau and J. Nathan Kutz and Steven L. Brunton},
+  title = {PySINDy: A comprehensive Python package for robust sparse system identification},
+  journal = {Journal of Open Source Software}
+  }
 @book{Brunton2019book,
   title={Data-driven science and engineering: Machine learning, dynamical systems, and control},
   author={{Brunton}, Steven L. and Kutz, J Nathan},
@@ -61,7 +89,17 @@
   doi={10.1017/9781108380690},
   publisher={Cambridge University Press}
 }
-
+@article{scikit-learn,
+ title={Scikit-learn: Machine Learning in {P}ython},
+ author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+         and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
+         and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
+         Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+ journal={Journal of Machine Learning Research},
+ volume={12},
+ pages={2825--2830},
+ year={2011}
+}
 @article{brunton2022modern,
 author = {Brunton, Steven L. and Budi\v{s}i\'{c}, Marko and Kaiser, Eurika and Kutz, J Nathan},
 title = {Modern {K}oopman Theory for Dynamical Systems},
