+++ .github/workflows/run-tests.yml
@@ -8,10 +8,10 @@ jobs:
 
     steps:
       - uses: actions/checkout@v1
-      - name: Set up Python 3.10.12
+      - name: Set up Python 3.10.13
         uses: actions/setup-python@v3
         with:
-          python-version: 3.10.12
+          python-version: 3.10.13
       - name: Linting
         run: |
           pip install pre-commit
@@ -22,7 +22,7 @@ jobs:
     strategy:
       max-parallel: 8
       matrix:
-        python-version: [3.10.12]
+        python-version: [3.10.13]
 
     steps:
     - uses: actions/checkout@v1
