+++ .github/workflows/release.yaml
@@ -15,7 +15,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v2
         with:
-          python-version: 3.10.12
+          python-version: 3.10.13
 
       - name: Install dependencies
         run: python -m pip install wheel twine
