### 2023-10-05 #5881 @pswpswpsw
Hi @fandreuz, thank you for your time reviewing! I just finished the revised paper. Please take a look.

1. Figures are now integrated with the manuscript.
2. Major dependency are justified with a separate paragraph in Feature section.
3. I have included a Example section to demonstrate a typical workflow using this package.
4. I have included comparison with pydmd in Statement of need. So basically our goal is to strike a balance between serving dynamics/control community and fluid dynamics community. Many of the unique designs/implementation of pykoopman is indeed to achieve that goal.
