### 2023-10-05 #5881 @fandreuz
Hi @pswpswpsw, thanks for updating the paper:

1. Fig2/3/4 are fine, but still I think Fig.1 is a little bit detached from the context. And since it takes half of a page, I would expect it to be either very self-explanatory, or to find an explanation in the body of the manuscript. To me it's confusing and not adding much value to the paper if it's presented this way, therefore I would spend some more words either in the caption or in the body.
2. Thanks for addressing my request, but I feel now it's too much. Of course you don't need to justify why you use each dependency, but I expected one, max two, sentences to explain your architecture. Example: *We use several infrastructure objects and utility functions from PyDMD, however we decided to implement our version of DMDs because ...*
3. I don't see how your package comes into play in your example code, am I missing something?
4. Thanks.

There's a typo/rendering problem at L93. I propose to remove "the" before "PyKoopman" in L51.

About *Statement of need*: please try to condense it. I propose to have a detailed explanation of all features supported in PyKoopman only in one place, namely the *Features* section. In *Statement of need* I'd like to see only 1/2 sentences about high-level features of your package, 1/2 sentences for comparison with state of the art (e.g. other open source packages like PyDMD) and much less citations. I would say there's not much need for many citations in *Statement of need* as anyway you will go into details in *Features*, so I propose to keep only the critical ones.
