+++ docs/JOSS/paper.md
@@ -1,258 +1,165 @@
 ---
-title: 'PyKoopman: A Python Package for Data-Driven Approximation of the Koopman
-Operator'
+title: 'PyKoopman: A Python Package for Data-Driven Approximation of the Koopman Operator'
 tags:
-
-- Python
-- dynamical systems
-- Koopman operator
-- system identification
-- machine learning
-- neural networks
-  authors:
-- name: Shaowu Pan
-  orcid: 0000-0002-2462-362X
-  affiliation: "1, 3"
-- name: Eurika Kaiser
-  orcid: 0000-0001-6049-0812
-  affiliation: 2
-- name: Brian M. de Silva
-  orcid: 0000-0003-0944-900X
-  affiliation: 1
-- name: J. Nathan Kutz
-  orcid: 0000-0002-6004-2275
-  affiliation: 1
-- name: Steven L. Brunton
-  orcid: 0000-0002-6565-5118
-  affiliation: 2
-  affiliations:
-- name: Department of Applied Mathematics, University of Washington
-  index: 1
-- name: Department of Mechanical Engineering, University of Washington
-  index: 2
-- name: Department of Mechanical, Aerospace, and Nuclear Engineering, Rensselaer
-  Polytechnic Institute
-  index: 3
-  date: 20 June 2023
-  bibliography: paper.bib
-
+  - Python
+  - dynamical systems
+  - Koopman operator
+  - system identification
+  - machine learning
+  - neural networks
+authors:
+  - name: Shaowu Pan
+    orcid: 0000-0002-2462-362X
+    affiliation: "1, 3"
+  - name: Eurika Kaiser
+    orcid: 0000-0001-6049-0812
+    affiliation: 2
+  - name: Brian M. de Silva
+    orcid: 0000-0003-0944-900X
+    affiliation: 1
+  - name: J. Nathan Kutz
+    orcid: 0000-0002-6004-2275
+    affiliation: 1
+  - name: Steven L. Brunton
+    orcid: 0000-0002-6565-5118
+    affiliation: 2
+affiliations:
+ - name: Department of Applied Mathematics, University of Washington
+   index: 1
+ - name: Department of Mechanical Engineering, University of Washington
+   index: 2
+ - name: Department of Mechanical, Aerospace, and Nuclear Engineering, Rensselaer Polytechnic Institute
+   index: 3
+date: 20 June 2023
+bibliography: paper.bib
 ---
 
 # Summary
 
-`PyKoopman` is a Python package for the data-driven approximation of the Koopman
-operator associated with a dynamical systems. The Koopman operator is a principled
-linear embedding of nonlinear dynamics and facilitates the prediction, estimation, and
-control of strongly nonlinear dynamics using linear systems theory. In
-particular, `PyKoopman` provides tools for data-driven system identification for
-unforced and actuated systems that build on the equation-free dynamic mode
-decomposition (DMD) [@schmid2010jfm] and its
-variants [@Kutz2016book;@schmid2022dynamic;@brunton2022modern]. In this work, we provide
-a brief description of the mathematical underpinnings of the Koopman operator, an
-overview and demonstration of the features implemented in `PyKoopman` (with code
-examples), practical advice for users, and a list of potential extensions to `PyKoopman`
-. Software is available at \url{https://github.com/dynamicslab/pyKoopman}.
[…]
