### 2023-10-06 #5881 @pswpswpsw
Hi @fandreuz,
Thanks for the second round review.

1. I have modified the caption to make it more concise. The figure 1 is to summarize the fundamental concepts in Koopman operator and also gives an example for why it can work.
2. I believe this information is helpful for the readers in case they ask. So I leave it there in case the second reviewer is interested to know.
3. Oh, I missed the rest of the code for `pykoopman`. Now I have added them to the revised version. Please take a look.

I just corrected the typo/rendering issues and also removed the "the" before Pykoopman.

For statement of need, the revised version now only has one sentence about comparison to `pydmd`:

> Compared to implementation of DMD (e.g., `PyDMD` [@demo2018pydmd]) which can be viewed as a linear projection of Koopman operator, `PyKoopman` offers a comprehensive set of _nonlinear_ projection methods.
