### 2023-10-27 #5881 @ulf1
@pswpswpsw has the package `PyKoopman` been used in academic papers?
