### 2023-10-30 25c7a17
fix issues with migration from setuptools to toml
