### 2023-10-30 bc6092a
fix issues with migration from setuptools to toml
