### 2023-10-30 030d396
fix issues with migration from setuptools to toml
