+++ requirements.txt
@@ -1,5 +1,3 @@
--e .
--r requirements-torch.txt --extra-index-url https://download.pytorch.org/whl/cu118/
 matplotlib == 3.5.0
 derivative
 scikit-learn < 1.2
