### 2023-10-30 a3839d3
fix issues with migration from setuptools to toml
