+++ .github/workflows/run-tests.yml
@@ -32,7 +32,8 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        pip install -r requirements-dev.txt
+        pip install -e .[dev]
+#        pip install -r requirements-dev.txt
     - name: Test with pytest
       run: |
         py.test test
