+++ pyproject.toml
@@ -1,19 +1,13 @@
-[build-system]
-requires = [
-    "setuptools>=61.0",
-    "wheel"]
-build-backend = "setuptools.build_meta"
-
 [project]
 dynamic = ["dependencies"]
 name = "pykoopman"
 version = "1.0.5"
 authors = [
-    { name="Shaowu Pan", email="shawnpan@uw.edu" },
-    { name="Eurika Kaiser", email="eurika@uw.edu" },
-    { name="Brian de Silva", email="bdesilva@uw.edu" },
-    { name="J. Nathan Kutz", email="kutz@uw.edu" },
-    { name="Steven L. Brunton", email="sbrunton@uw.edu" },
+    { name = "Shaowu Pan", email = "shawnpan@uw.edu" },
+    { name = "Eurika Kaiser", email = "eurika@uw.edu" },
+    { name = "Brian de Silva", email = "bdesilva@uw.edu" },
+    { name = "J. Nathan Kutz", email = "kutz@uw.edu" },
+    { name = "Steven L. Brunton", email = "sbrunton@uw.edu" },
 ]
 description = "Python package for data-driven approximations to the Koopman operator."
 readme = "README.rst"
@@ -26,10 +20,51 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Topic :: Scientific/Engineering :: Mathematics",
 ]
+dependencies = [
+    "matplotlib == 3.5.0",
+    "derivative",
+    "scikit-learn < 1.2",
+    "numpy",
+    "scipy",
+    "pydmd",
+    "optht",
+    "scipy",
+    "prettytable",
+    "torch",
+    "torchvision",
+    "lightning"
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest-cov",
+    "pytest-lazy-fixture",
+    "flake8-builtins-unleashed",
+    "setuptools_scm",
+    "setuptools_scm_git_archive",
+    "jupyter",
+    "notebook",
+    "nbsphinx",
+    "sphinx-codeautolink",
+    "sphinx >= 2,<7.0.0",
+    "sphinxcontrib-apidoc",
+    "sphinx_rtd_theme",
+    "pre-commit",
+    "sphinx-nbexamples",
+    "jupyter_contrib_nbextensions",
+    "PyQt5",
+    "osqp",
+]
+
+[build-system]
+requires = [
+    "setuptools>=61.0",
+    "wheel"]
+build-backend = "setuptools.build_meta"
 
 [project.urls]
 "Homepage" = "https://github.com/dynamicslab/pykoopman"
 "Bug Tracker" = "https://github.com/dynamicslab/pykoopman/issues"
 
 [tool.setuptools.dynamic]
-dependencies = {file = ["requirements.txt"]}
+dependencies = { file = ["requirements.txt"] }
