### 2023-10-30 9c815fe
fix issues with migration from setuptools to toml
