### 2023-10-30 d5a442a
fix issues with migration from setuptools to toml
