+++ .readthedocs.yaml
@@ -24,8 +24,8 @@ sphinx:
 #    - pdf
 
 # Optionally declare the Python requirements required to build your docs
-python:
-   install:
-    - requirements: requirements-dev.txt
-    - method: pip
-      path: .
+#python:
+#   install:
+#    - requirements: requirements-dev.txt
+#    - method: pip
+#      path: .
