### 2023-10-31 871f7f4
fix issues with migration from setuptools to toml
