### 2023-10-31 1540af4
fix issues with migration from setuptools to toml
