### 2023-10-31 fd5c510
fix issues with migration from setuptools to toml
