+++ requirements.txt
@@ -1,10 +1,0 @@
-matplotlib == 3.5.0
-derivative
-scikit-learn < 1.2
-numpy
-scipy
-# pydmd @ git+https://github.com/mathLab/PyDMD@c98ede0#egg=pydmd
-pydmd
-optht
-scipy
-prettytable
