+++ pyproject.toml
@@ -67,5 +67,6 @@ dev = [
 "Homepage" = "https://github.com/dynamicslab/pykoopman"
 "Bug Tracker" = "https://github.com/dynamicslab/pykoopman/issues"
 
+[tool.setuptools_scm]
 #[tool.setuptools.dynamic]
 #dependencies = { file = ["requirements.txt"] }
