### 2023-10-31 f88092d
fix issues with migration from setuptools to toml
