### 2023-10-31 c99bd1e
fix issues with migration from setuptools to toml
