### 2023-10-31 6bc5ff9
fix issues with migration from setuptools to toml
