+++ pyproject.toml
@@ -16,7 +16,7 @@ authors = [
 ]
 description = "Python package for data-driven approximations to the Koopman operator."
 readme = "README.rst"
-requires-python = ">=3.10"
+requires-python = ">=3.10, <3.11"
 classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3.10",
@@ -26,18 +26,18 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Mathematics",
 ]
 dependencies = [
-    "matplotlib~=3.5.0",
-    "derivative",
-    "scikit-learn<1.2",
-    "numpy",
-    "scipy",
-    "pydmd",
-    "optht",
-    "scipy",
-    "prettytable",
-    "torch",
-    "torchvision",
-    "lightning",
+    "matplotlib >3.0, <= 3.5.0",
+    "derivative ~= 0.6.0",
+    "scikit-learn >= 1.1.3, <= 1.2",
+    "numpy >=1.20, <= 1.26",
+    "scipy >1.6.0, <= 1.11.2",
+    "pydmd >0.4, <= 0.4.1",
+    "optht ~= 0.2.0",
+    "prettytable >3.0.0, <= 3.9.0",
+    "torch > 2.0, <= 2.1.0",
+    "torchvision > 0.14, <= 0.16.0",
+    "torchaudio > 2.0, <= 2.1.0",
+    "lightning > 2.0, <= 2.0.9",
 ]
 
 [project.optional-dependencies]
@@ -47,18 +47,16 @@ dev = [
     "flake8-builtins-unleashed",
     "setuptools_scm",
     "setuptools_scm_git_archive",
-    "jupyter",
-    "notebook",
+    "jupyter >= 1.0.0",
+    "notebook >7.0.0, <=7.0.4",
     "nbsphinx",
     "sphinx-codeautolink",
-    "sphinx >= 2,<7.0.0",
+    "sphinx >= 3,<=6.2.1",
     "sphinxcontrib-apidoc",
     "sphinx_rtd_theme",
     "pre-commit",
     "sphinx-nbexamples",
     "jupyter_contrib_nbextensions",
-    "PyQt5",
-    "osqp",
 ]
 
 
