### 2023-10-31 #5881 @pswpswpsw
> [ulf1](/ulf1)
Hi Ulf, first I want to thank you for pointing out the issues. I learned a lot.

To answer your question, not yet (because we just officially released the paper) but soon I will have a few papers that uses this pykoopman. I believe the community will also start to use our package for benchmark.
