+++ pyproject.toml
@@ -42,10 +42,10 @@ dependencies = [
 
 [project.optional-dependencies]
 dev = [
-    "pytest-cov",
-    "pytest-lazy-fixture",
-    "flake8-builtins-unleashed",
-    "setuptools_scm",
+    "pytest-cov ~= 4.1.0",
+    "pytest-lazy-fixture ~= 0.6.3",
+    "flake8-builtins-unleashed ~= 1.3.1",
+    "setuptools_scm ~= 8.0.2",
     "setuptools_scm_git_archive",
     "jupyter >= 1.0.0",
     "notebook >7.0.0, <=7.0.4",
