+++ .github/workflows/run-tests.yml
@@ -36,7 +36,7 @@ jobs:
 #        pip install -r requirements-dev.txt
     - name: Test with pytest
       run: |
-        py.test test
+        py.test -v test
     - name: Generate coverage report
       run: |
         pip install pytest
