+++ pyproject.toml
@@ -16,7 +16,7 @@ authors = [
 ]
 description = "Python package for data-driven approximations to the Koopman operator."
 readme = "README.rst"
-requires-python = ">=3.10, <3.11"
+requires-python = ">=3.10"
 classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3.10",
@@ -26,7 +26,7 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Mathematics",
 ]
 dependencies = [
-    "matplotlib >3.0, <= 3.5.0",
+    "matplotlib ~= 3.5.0",
     "derivative ~= 0.6.0",
     "scikit-learn >= 1.1.3, <= 1.2",
     "numpy >=1.20, <= 1.26",
@@ -51,7 +51,7 @@ dev = [
     "notebook >7.0.0, <=7.0.4",
     "nbsphinx",
     "sphinx-codeautolink",
-    "sphinx >= 3,<=6.2.1",
+    "sphinx >= 3,<=7.0.0",
     "sphinxcontrib-apidoc",
     "sphinx_rtd_theme",
     "pre-commit",
