+++ src/pykoopman/regression/_havok.py
@@ -1,0 +1,327 @@
+"""module for havok"""
+from __future__ import annotations
+
+from warnings import warn
+
+import numpy as np
+from matplotlib import pyplot as plt
+from optht import optht
+from scipy.signal import lsim
+from scipy.signal import lti
+from sklearn.utils.validation import check_is_fitted
+
+from ..common import drop_nan_rows
+from ..differentiation._derivative import Derivative
+from ._base import BaseRegressor
+
+
+class HAVOK(BaseRegressor):
+    """
+    HAVOK (Hankel Alternative View of Koopman) regressor.
+
+    Aims to determine the system matrices A, B that satisfy d/dt v = Av + Bu,
+    where v is the vector of the leading delay coordinates and u is a low-energy
+    delay coordinate acting as forcing. A and B are the unknown system and control
+    matrices, respectively. The delay coordinates are obtained by computing the
+    SVD from a Hankel matrix.
+
+    The objective function, \\|dV-AV-BU\\|_F, is minimized using least-squares
+    regression.
+
+    See the following reference for more details:
+        Brunton, S.L., Brunton, B.W., Proctor, J.L., Kaiser, E. & Kutz, J.N.
+        "Chaos as an intermittently forced linear system."
+        Nature Communications, Vol. 8(19), 2017.
+        <https://www.nature.com/articles/s41467-017-00030-8>
+
+    Parameters:
+        svd_rank (int, optional):
+            Rank of the SVD used for model reduction. Defaults to None.
+        differentiator (Derivative, optional):
+            Differentiation method to compute the time derivative. Defaults to
+            Derivative(kind="finite_difference", k=1).
+        plot_sv (bool, optional):
+            Whether to plot the singular values. Defaults to False.
+
+    Attributes:
+        coef_ (array):
+            Weight vectors of the regression problem. Corresponds to either [A] or
+            [A,B].
+        state_matrix_ (array):
+            Identified state transition matrix A of the underlying system.
+        control_matrix_ (array):
+            Identified control matrix B of the underlying system.
+        projection_matrix_ (array):
+            Projection matrix into low-dimensional subspace of shape (n_input_features
+            +n_control_features, svd_rank).
+        projection_matrix_output_ (array):
+            Projection matrix into low-dimensional subspace of shape (n_input_features
+            +n_control_features, svd_output_rank).
+    """
+
+    def __init__(
+        self,
+        svd_rank=None,
+        differentiator=Derivative(kind="finite_difference", k=1),
+        plot_sv=False,
+    ):
+        """
+        Initialize the HAVOK regressor.
+
+        Args:
+            svd_rank (int, optional):
+                Rank of the SVD used for model reduction. Defaults to None.
+            differentiator (Derivative, optional):
+                Differentiation method to compute the time derivative. Defaults to
+                Derivative(kind="finite_difference", k=1).
+            plot_sv (bool, optional):
+                Whether to plot the singular values. Defaults to False.
+        """
+        self.svd_rank = svd_rank
+        self.differentiator = differentiator
+        self.plot_sv = plot_sv
+
+    def fit(self, x, y=None, dt=None):
+        """
+        Fit the HAVOK regressor to the given data.
+
+        Args:
+            x (numpy.ndarray):
+                Measurement data to be fit.
+            y (not used):
+                Time-shifted measurement data to be fit. Ignored.
+            dt (scalar):
+                Discrete time-step.
+
+        Returns:
+            self: Fitted HAVOK instance.
[…]
