+++ pyproject.toml
@@ -26,9 +26,9 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Mathematics",
 ]
 dependencies = [
-    "matplotlib ~= 3.5.0",
+    "matplotlib <= 3.5.0, >=3.5.0",
     "derivative ~= 0.6.0",
-    "scikit-learn >= 1.1.3, <= 1.2",
+    "scikit-learn >= 1.1.3, <= 1.1.3",
     "numpy >=1.20, <= 1.26",
     "scipy >1.6.0, <= 1.11.2",
     "pydmd >0.4, <= 0.4.1",
