### 2023-11-21 #5881 @fandreuz
Hi @pswpswpsw, as I said in my last comment I still feel the full detailed summary for each dependency is not needed and it should be cut. Not sure what's the opinion of @ulf1 here.

I also feel 11 pages is quite a lot for JOSS, in general I would:
- Cut some references, it's not interesting to list papers treating similar topics, I would restrict to no more than 2/3 per sub-topic and move the extended list to your repository;
- Cut dependencies as said above
- Shorten the example. No need for import statements, generation of dataset and plotting. Just put what makes your library peculiar, so the reader is interested in having a look at the examples in your repository.

I've checked all the other points.
