### 2023-11-24 #5881 @ulf1
@olexandr-konovalov  @pswpswpsw @fandreuz Hello, I finished my review. I like the jupyter demo notebooks to explain how to call the package's classes, and the unit tests. The dependencies, including version ranges, have been fixed.
I agree with @fandreuz that the PDF is very lengthy.
I recommend this submission for publication.
