### 2023-11-26 #5881 @pswpswpsw
Hi @fandreuz , thanks for the reply. I have
- shorten length to 8 pages.
- removed a lot of discussion on similar topics
- removed detailed discussion on dependency

Hope this would work. Thanks.
