### 2024-01-28 #5881 @olexandr-konovalov
Thanks @fandreuz and @ulf1 - yes, @pswpswpsw, I have now to follow [this process](https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance), let me try to find some time next week...
