### 2024-02-07 fe992d4
fix bug in prediction and upgrade nndmd to handle list of trajectories even at the level of regressor
