+++ pyproject.toml
@@ -42,6 +42,7 @@ dependencies = [
 
 [project.optional-dependencies]
 dev = [
+    "pytest <= 7.4.4",
     "pytest-cov ~= 4.1.0",
     "pytest-lazy-fixture ~= 0.6.3",
     "flake8-builtins-unleashed ~= 1.3.1",
