+++ test/regression/test_regressors.py
@@ -10,6 +10,7 @@ from pykoopman.regression import KDMD
 from pykoopman.regression import NNDMD
 from pykoopman.regression import PyDMDRegressor
 from sklearn.gaussian_process.kernels import RBF
+import pykoopman as pk
 
 
 class RegressorWithoutFit:
@@ -70,16 +71,26 @@ def test_fit_regressors(data_xy, regressor):
     [
         # case 1,2 only work for pykoopman class
         # case 1: single step single traj, no validation
-        (np.random.rand(200, 3), None),
+        (
+                np.random.rand(200, 3),
+                None
+        ),
         # case 2: single step multiple traj, no validation
-        (np.random.rand(200, 3), np.random.rand(200, 3)),
+        (
+                np.random.rand(200, 3),
+                np.random.rand(200, 3) # because "x" is not a list, so we think this
+                                       # is single step
+        ),
         # case 3,4 works for regressor directly
         # case 3: multiple traj, no validation
-        ([np.random.rand(200, 3), np.random.rand(100, 3)], None),
+        (
+                [np.random.rand(200, 3), np.random.rand(100, 3)],  # this is training
+                None  # no validation
+        ),
         # case 4: multiple traj, with validation
         (
-            [np.random.rand(100, 3), np.random.rand(100, 3)],
-            [np.random.rand(300, 3), np.random.rand(400, 3)],
+            [np.random.rand(100, 3), np.random.rand(100, 3)], # this is training
+            [np.random.rand(300, 3), np.random.rand(400, 3)], # this is validation
         ),
     ],
 )
@@ -107,3 +118,56 @@ def test_fit_nndmd_regressor(data_xy, regressor):
     """test if using nndmd regressor alone will run the fit without error"""
     x, y = data_xy
     regressor.fit(x, y)
+
+@pytest.mark.parametrize(
+    "data_xy",
+    [
+        # # case 1,2 only work for pykoopman class
+        # # case 1: single step single traj, no validation
+        # (
+        #         np.random.rand(200, 3),
+        #         None
+        # ),
+        # # case 2: single step multiple traj, no validation
+        # (
+        #         np.random.rand(200, 3),
+        #         np.random.rand(200, 3) # because "x" is not a list, so we think this
+        #                                # is single step
+        # ),
+        # # case 3,4 works for regressor directly
+        # # case 3: multiple traj, no validation
+        # (
+        #         [np.random.rand(200, 3), np.random.rand(100, 3)],  # this is training
+        #         None  # no validation
+        # ),
+        # case 4: multiple traj, with validation
+        (
+            [np.random.rand(100, 3), np.random.rand(100, 3)], # this is training
+            [np.random.rand(300, 3), np.random.rand(400, 3)], # this is validation
+        ),
+    ],
+)
+@pytest.mark.parametrize(
+    "regressor",
+    [
+        NNDMD(
+            mode="Dissipative",
+            look_forward=2,
+            config_encoder=dict(
+                input_size=3, hidden_sizes=[32] * 2, output_size=4, activations="swish"
+            ),
+            config_decoder=dict(
+                input_size=4, hidden_sizes=[32] * 2, output_size=3, activations="linear"
+            ),
+            batch_size=512,
+            lbfgs=True,
+            normalize=False,
+            normalize_mode="max",
+            trainer_kwargs=dict(max_epochs=1),
+        )
+    ],
+)
+def test_fit_dlkoopman(data_xy, regressor):
+    """test if using NNDMD regressor work inside pykoopman"""
+    model_d = pk.Koopman(regressor=regressor)
+    model_d.fit(data_xy[0],data_xy[1], dt=1)
