+++ docs/JOSS/paper.bib
@@ -68,6 +68,7 @@
   volume={3},
   number={22},
   pages={530},
+  doi={10.21105/joss.00530},
   year={2018}
 }
 @article{kaptanoglu2022pysindy,
