### 2024-02-18 #5881 @pswpswpsw
@olexandr-konovalov Hi Editor, I have ensured our page on zenodo (https://zenodo.org/records/10633346) has the correct author list and ORCID. I have completed the following.
- Note that the work is performed when the third author was in the previous insititution.
- We have made 1.0.9 as our latest version.
- We have archive the release on zenodo.
- The title and author list matches between Zenodo and the JOSS paper
- The liscence is also the same.
