### 2024-02-19 #5881 @olexandr-konovalov
@pswpswpsw thank you - I've printed and proofread it (sorry took a while to reach it), and I like the current shape of the article, but here there are some, hopefully final, suggestions:
- [ ] page 2, figure 1: if you still have an editable version of the figure, would you agree that it will be useful to add equality sign after `z`, writing `x = [...] z =` in the 1st row of the formula? It seems to me that a hanging "=" would help to indicate a continuation.
- [ ] page 3, Fig. 2: only if you have an editable source - the misalignment of PyDMD and PyTorch in the bottom row would be nice to fix.
- [ ] page 3, lines 55-58: you name several Python packages not using spelling and capitalisations from their official names, but using names of the respective Python packages, e.g. `PyDMD` vs `pydmd`, `scikit-learn` vs `sklearn`, `pytorch` vs `torch`. I suggest to stick to the official names of packages, and also to add proper software citations for those listed in this paragraph.
- [ ] page 3, line 69 .. page 4, line 109: the style of presenting things in the 1st block differs from the rest, and this is confusing. I suggest to the style used in "System identification method.." and below, e.g. "- what is implemented (reference): `where_it_is implemented`".
- [ ] page 5, line 121: mismatched quotes around "Koopman" (should it be in monospace font like `Koopman` referring to an object from code block above?)
- [ ] page 7, lines 202 and 221: Koopman should start in uppercase
