+++ docs/JOSS/paper.bib
@@ -20,8 +20,44 @@
   doi={10.1137/1.9781611974508},
   publisher={SIAM}
 }
-
-
+@software{Falcon_PyTorch_Lightning_2019,
+author = {Falcon, William and {The PyTorch Lightning team}},
+doi = {10.5281/zenodo.3828935},
+license = {Apache-2.0},
+month = mar,
+title = {{PyTorch Lightning}},
+url = {https://github.com/Lightning-AI/lightning},
+version = {1.4},
+year = {2019}
+}
+@article{kaptanoglu2022pysindy,
+  doi = {10.21105/joss.03994},
+  url = {https://doi.org/10.21105/joss.03994},
+  year = {2022},
+  publisher = {The Open Journal},
+  volume = {7},
+  number = {69},
+  pages = {3994},
+  author = {Alan A. Kaptanoglu and Brian M. de Silva and Urban Fasel and Kadierdan Kaheman and Andy J. Goldschmidt and Jared Callaham and Charles B. Delahunt and Zachary G. Nicolaou and Kathleen Champion and Jean-Christophe Loiseau and J. Nathan Kutz and Steven L. Brunton},
+  title = {PySINDy: A comprehensive Python package for robust sparse system identification},
+  journal = {Journal of Open Source Software}
+  }
+@article{pedregosa2011scikit,
+  title={Scikit-learn: Machine learning in Python},
+  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
+  journal={the Journal of machine Learning research},
+  volume={12},
+  pages={2825--2830},
+  year={2011},
+  publisher={JMLR. org}
+}
+@article{paszke2019pytorch,
+  title={Pytorch: An imperative style, high-performance deep learning library},
+  author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
+  journal={Advances in neural information processing systems},
+  volume={32},
+  year={2019}
+}
 @article{schmid2022dynamic,
 	author = {Schmid, Peter J},
 	journal = {Annual Review of Fluid Mechanics},
@@ -549,7 +585,7 @@ keywords = {Data-driven scientific computing, Machine learning, Predictive model
 	journal = {SIAM Journal on Applied Dynamical Systems},
 	number = {3},
 	pages = {2162--2193},
-	title = {Data-driven model predictive control using interpolated Koopman generators},
+	title = {Data-driven model predictive control using interpolated {K}oopman generators},
 	volume = {19},
   doi={10.1137/20M1325678},
 	year = {2020}}
