+++ .github/workflows/release.yaml
@@ -18,7 +18,7 @@ jobs:
           python-version: 3.10.13
 
       - name: Install dependencies
-        run: python -m pip install wheel twine
+        run: python -m pip install wheel twine build
 
       - name: Build package
         run: python -m build
