+++ .github/workflows/release.yaml
@@ -1,30 +1,0 @@
-name: Release
-on:
-  release:
-    types:
-      - published
-
-jobs:
-  release:
-    name: Deploy release to PyPI
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout source
-        uses: actions/checkout@v2
-
-      - name: Set up Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: 3.10.13
-
-      - name: Install dependencies
-        run: python -m pip install wheel twine build
-
-      - name: Build package
-        run: python -m build
-
-      - name: Upload package
-        uses: pypa/gh-action-pypi-publish@master
-        with:
-          user: __token__
-          password: ${{ secrets.PYPI_TOKEN }}
