### 2024-02-20 #5881 @olexandr-konovalov
@pswpswpsw thanks! I hope we can now progress quickly.

One more instance of `koopman` starting in lowercase - page 8, line 236.

When fixed, please make a new release and archive it on Zenodo, and let me know.

Archiving on Zenodo, you may have to edit your archive metadata, to change "dynamicslab/pykoopman: v1.0.9" to "PyKoopman: A Python Package for Data-Driven Approximation of the Koopman Operator" to match the title of the paper - compare how this is done in any of the recently published JOSS papers.
