### 2024-02-20 #5881 @olexandr-konovalov
@pswpswpsw actually, a new release is only required if additional changes to the software were made - you don't have to make a release, but just update the metadata, to make the title matching (thanks @kyleniemeyer for the clarification)
