+++ docs/JOSS/paper.bib
@@ -517,7 +517,7 @@ keywords = {Data-driven scientific computing, Machine learning, Predictive model
 	booktitle = {Advances in Neural Information Processing Systems},
 	pages = {1130--1140},
   doi={10.48550/arXiv.1710.04340},
-	title = {Learning Koopman Invariant Subspaces for Dynamic Mode Decomposition},
+	title = {Learning {K}oopman Invariant Subspaces for Dynamic Mode Decomposition},
 	year = {2017}}
 
 @article{pan2021sparsity,
