### 2024-02-20 #5881 @olexandr-konovalov
@pswpswpsw after all, we might need a new release - https://github.com/dynamicslab/pykoopman/tree/v1.0.9 has a lot of files ending with `~` which seem to be added accidentally. Sorry, haven't noticed before!
