### 2024-02-20 #5881 @olexandr-konovalov
The repository looks good. Next, https://zenodo.org/records/10685233 says v1.1.0, but surprisingly it sats in two places

- Is supplement to Software: [https://github.com/dynamicslab/pykoopman/tree/v1.0.9 ](https://github.com/dynamicslab/pykoopman/tree/v1.0.9)(URL)

- External resources: Available in [dynamicslab/pykoopman](https://github.com/dynamicslab/pykoopman/tree/v1.0.9) Release: v1.0.9

Have you set up everything like https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content suggests? If so, that's puzzling to see. Anyhow, perhaps this is something you can fix on Zenodo manually?
