### 2024-02-20 #5881 @olexandr-konovalov
@pswpswpsw I am investigating this - can't see any obvious reason for this, looking at references.
