### 2024-02-20 #5881 @olexandr-konovalov
@pswpswpsw aaah, now it makes sense - look at https://github.com/dynamicslab/pykoopman/blob/master/docs/JOSS/paper.bib, it has duplicate labels `paszke2019pytorch` and `kaptanoglu2022pysindy`. Please fix!
