+++ docs/JOSS/paper.bib
@@ -51,13 +51,7 @@ year = {2019}
   year={2011},
   publisher={JMLR. org}
 }
-@article{paszke2019pytorch,
-  title={Pytorch: An imperative style, high-performance deep learning library},
-  author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
-  journal={Advances in neural information processing systems},
-  volume={32},
-  year={2019}
-}
+
 @article{schmid2022dynamic,
 	author = {Schmid, Peter J},
 	journal = {Annual Review of Fluid Mechanics},
@@ -107,18 +101,7 @@ year = {2019}
   doi={10.21105/joss.00530},
   year={2018}
 }
-@article{kaptanoglu2022pysindy,
-  doi = {10.21105/joss.03994},
-  url = {https://doi.org/10.21105/joss.03994},
-  year = {2022},
-  publisher = {The Open Journal},
-  volume = {7},
-  number = {69},
-  pages = {3994},
-  author = {Alan A. Kaptanoglu and Brian M. de Silva and Urban Fasel and Kadierdan Kaheman and Andy J. Goldschmidt and Jared Callaham and Charles B. Delahunt and Zachary G. Nicolaou and Kathleen Champion and Jean-Christophe Loiseau and J. Nathan Kutz and Steven L. Brunton},
-  title = {PySINDy: A comprehensive Python package for robust sparse system identification},
-  journal = {Journal of Open Source Software}
-  }
+
 @book{Brunton2019book,
   title={Data-driven science and engineering: Machine learning, dynamical systems, and control},
   author={{Brunton}, Steven L. and Kutz, J Nathan},
