### 2024-02-21 #5881 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
