+++ docs/JOSS/paper.bib
@@ -39,11 +39,11 @@ year = {2019}
   number = {69},
   pages = {3994},
   author = {Alan A. Kaptanoglu and Brian M. de Silva and Urban Fasel and Kadierdan Kaheman and Andy J. Goldschmidt and Jared Callaham and Charles B. Delahunt and Zachary G. Nicolaou and Kathleen Champion and Jean-Christophe Loiseau and J. Nathan Kutz and Steven L. Brunton},
-  title = {PySINDy: A comprehensive Python package for robust sparse system identification},
+  title = {PySINDy: A comprehensive {P}ython package for robust sparse system identification},
   journal = {Journal of Open Source Software}
   }
 @article{pedregosa2011scikit,
-  title={Scikit-learn: Machine learning in Python},
+  title={Scikit-learn: Machine learning in {P}ython},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
   journal={the Journal of machine Learning research},
   volume={12},
@@ -582,7 +582,7 @@ keywords = {Data-driven scientific computing, Machine learning, Predictive model
   number = {49},
   pages = {2104},
   author = {Brian M. de Silva and Kathleen Champion and Markus Quade and Jean-Christophe Loiseau and J. Nathan Kutz and Steven L. Brunton},
-  title = {PySINDy: A Python package for the sparse identification of nonlinear dynamical systems from data},
+  title = {PySINDy: A {P}ython package for the sparse identification of nonlinear dynamical systems from data},
   journal = {Journal of Open Source Software}
 }
 
@@ -642,7 +642,7 @@ keywords = {Data-driven scientific computing, Machine learning, Predictive model
 	journal = {Machine Learning: Science and Technology},
 	number = {1},
 	pages = {015009},
-	title = {Deeptime: a Python library for machine learning dynamical models from time series data},
+	title = {Deeptime: a {P}ython library for machine learning dynamical models from time series data},
 	volume = {3},
   doi={10.1088/2632-2153/ac3de0},
 	year = {2021}}
