### 2024-02-24 #5881 @danielskatz
👋 @pswpswpsw - please add countries to the affiliations in the paper. In addition, I've suggested some small changes in https://github.com/dynamicslab/pykoopman/pull/46  Please let me know when these two items are done (or if you disagree with any of them), and we can proceed.
